[refactor]: CalculatorItemQueue 더블 스택 큐로 변경
연결 리스트로 구현되어있던 Queue 구조를 더블 스택 큐로 변경
